Re-structuring, button actions, logo added
I did a bit of re-structuring to the folder structure.
Included a utility function to better resolve file paths.
Made the yes and no buttons disappear when the last page of a questionnaire is reached.
A low-opacity version of the Pet IQ Logo is now the app background, which required an implementation of the QResource system.